backend/requirements.txt
django
django-ninja
django-ninja-extra
django-cors-headers
djangorestframework